requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
psycopg2-binary==2.9.9
httpx==0.27.0
starlette==0.36.3
